+++ blockchain-challenge/src/components/features/NetworkStatus.tsx
